requirements.txt
aequilibrae==0.7.3
attrs==21.4.0
autopep8==1.6.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
click-default-group-wheel==1.2.2
click-plugins==1.1.1
cligj==0.7.2
cycler==0.11.0
Cython==0.29.30
Fiona==1.8.21
fonttools==4.33.3
geographiclib==1.52
geopandas==0.10.2
geopy==2.2.0
graph-articulations==0.0.1
graphs==0.1.3
graphviz==0.20
idna==3.3
joblib==1.1.0
kiwisolver==1.4.3
matplotlib==3.5.2
munch==2.5.0
numpy==1.22.4
packaging==21.3
pandas==1.4.2
Pillow==9.1.1
pyaml==21.10.1
pyarrow==8.0.0
pycodestyle==2.8.0
pyparsing==3.0.9
pyproj==3.3.1
pyshp==2.3.0
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.28.0
Rtree==1.0.0
scikit-learn==1.1.1
scipy==1.8.1
shapefile-to-GRP==0.0.4
shapefile-to-sqlite==0.4.1
Shapely==1.8.2
six==1.16.0
sqlite-fts4==1.0.1
sqlite-utils==3.26.1
tabulate==0.8.9
threadpoolctl==3.1.0
toml==0.10.2
urllib3==1.26.9
